Fix in-place multiplication for numpy 1.10.4

+++ combinato/extract/tools.py
@@ -62,12 +62,12 @@ class ExtractNcsFile(object):
         read data from an ncs file
         """
         data, times = self.ncs_file.read(start, stop, 'both')
-        fdata = np.array(data)
+        fdata = np.array(data).astype(np.float32)
         fdata *= (1e6 * self.ncs_file.header['ADBitVolts'])
 
         if self.ref_file is not None:
             ref_data = self.ref_file.read(start, stop, 'data')
-            fref_data = np.array(ref_data)
+            fref_data = np.array(ref_data).astype(np.float32)
             fref_data *= 1e6 * self.ref_file.header['ADBitVolts']
             fdata -= fref_data
